refactor
separate firebase class into 3 new classes
