+++ src/components/Modal/AddItemModal.tsx
@@ -1,7 +1,8 @@
 import Modal from "react-modal";
 import React from "react";
 import TextInput from "../Inputs/TextInput";
-import { Firebase } from "../../services/firebase";
+import { AuthManager } from "../../services/AuthManager";
+import { FirestoreManager } from "../../services/FirestoreManager";
 import { ItemEntity } from "../../entity/item-entity";
 import "../../style/AddItemModal.css"
 
@@ -16,13 +17,13 @@ interface FormParams {
 
 function AddItemModal() {
   const [modalIsOpen, setModalIsOpen] = React.useState(false);
-  const auth = Firebase.GetAuth();
+  const auth = AuthManager.GetAuth();
 
   async function handleSubmit(event: React.SyntheticEvent<HTMLFormElement>) {
     event.preventDefault();
     const form = event.currentTarget;
     const formElements = form.elements as typeof form.elements & FormParams;
-    await Firebase.AddProduct({
+    await FirestoreManager.AddProduct({
       title: formElements.title.value,
       description: formElements.description.value,
       email: auth.currentUser?.email,
