+++ src/pages/Product List/ProductListView.tsx
@@ -1,7 +1,7 @@
 import ProductView from "../../components/ProductView";
 import "../../style/ProductListView.css";
 import Product from "../../interfaces/Product";
-import { Firebase } from "../../services/firebase";
+import { FirestoreManager } from "../../services/FirestoreManager";
 import { useState, useEffect } from "react";
 
 const getFilteredItems = (query: string, items: Product[]) => {
@@ -23,7 +23,7 @@ function ProductListView() {
   const filteredItems = getFilteredItems(query, minhaLista);
 
   useEffect(() => {
-    Firebase.getProducts().then((productList) => setMinhaLista(productList));
+    FirestoreManager.getProducts().then((productList) => setMinhaLista(productList));
   }, []);
 
   const listItems = filteredItems.map((product) => (
